+++ rust-src/concordium_base/src/lib.rs
@@ -32,6 +32,12 @@ pub mod ps_sig;
 
 pub mod dodis_yampolskiy_prf;
 
+/// We expose the `PublicKey`, `SecretKey`, and `Signature` from the third-party
+/// `ed25519_dalek` crate here because these types appear in Concordium's API.
+pub mod ed25519 {
+    pub use ed25519_dalek::{PublicKey, SecretKey, Signature};
+}
+
 #[cfg(feature = "ffi")]
 mod ffi_helpers;