+++ rust-src/concordium_base/CHANGELOG.md
@@ -1,6 +1,13 @@
 ## Unreleased changes
 
+## 3.2.0 (2023-11-22)
+
+- Add `From` trait to convert `AccountKeys` into `AccountPublicKeys`.
+- Add `singleton` and `new` function to `AccountAccessStructure`.
+- Export `PublicKey`, `SecretKey`, and `Signature` type from `ed25519_dalek` crate.
 - Add `sign_message` function to sign a message with all `AccountKeys`. The return type is `AccountSignatures`.
+- Support using `validatorId` instead of `bakerId` when parsing
+  `BakerCredentials` from JSON.
 
 ## 3.1.1 (2023-10-27)